backend: timezone handling added, monthly values for /total-income-expense

+++ server/models/transactions.model.js
@@ -4,7 +4,7 @@ import sequelize from "../config/db.js";
 const Transactions = sequelize.define("transactions",
     {
         description: DataTypes.STRING(255),
-        amount: DataTypes.DOUBLE(10, 4),
+        amount: DataTypes.DOUBLE(10, 2),
         category: DataTypes.TINYINT,
         type: DataTypes.STRING(10)
